chore: member, group을 domain모듈로 통합
