+++ subprojects/domain/src/main/java/com/money/domain/group/entity/Anniversary.java
@@ -1,0 +1,19 @@
+package com.money.domain.group.entity;
+
+import jakarta.persistence.Embeddable;
+import java.time.LocalDate;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Embeddable
+public class Anniversary {
+    private LocalDate anniversaryDate;
+    private String anniversaryName;
+}
