+++ subprojects/domain/src/main/java/com/money/domain/group/entity/Group.java
@@ -1,0 +1,37 @@
+package com.money.domain.group.entity;
+
+import com.money.domain.BaseEntity;
+import com.money.domain.member.entity.Member;
+import jakarta.persistence.Embedded;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
+import java.util.ArrayList;
+import java.util.List;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Group extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String inviteCode;
+    private Long memberCount;
+
+    @OneToMany(mappedBy = "member_id")
+    List<Member> members = new ArrayList<>();
+
+    @Embedded
+    private Anniversary anniversary;
+}
