+++ subprojects/domain/src/main/java/com/money/domain/member/exception/ExpiredCodeException.java
@@ -1,0 +1,11 @@
+package com.money.domain.member.exception;
+
+import com.money.exception.ErrorCode;
+import com.money.exception.MainException;
+
+public class ExpiredCodeException extends MainException {
+
+    public ExpiredCodeException() {
+        super(ErrorCode.EXPIRED_CODE);
+    }
+}
