+++ subprojects/domain/src/main/java/com/money/domain/member/exception/MemberAlreadyExistException.java
@@ -1,0 +1,11 @@
+package com.money.domain.member.exception;
+
+import com.money.exception.ErrorCode;
+import com.money.exception.MainException;
+
+public class MemberAlreadyExistException extends MainException {
+
+    public MemberAlreadyExistException() {
+        super(ErrorCode.ALREADY_EXIST_MEMBER);
+    }
+}
