+++ subprojects/domain/src/main/java/com/money/domain/member/exception/NotFoundMemberException.java
@@ -1,0 +1,11 @@
+package com.money.domain.member.exception;
+
+import com.money.exception.ErrorCode;
+import com.money.exception.MainException;
+
+public class NotFoundMemberException extends MainException {
+
+    public NotFoundMemberException() {
+        super(ErrorCode.NOT_FOUND_MEMBER);
+    }
+}
