+++ subprojects/domain/src/main/java/com/money/domain/member/exception/OverFlowInviteCodeCountException.java
@@ -1,0 +1,11 @@
+package com.money.domain.member.exception;
+
+import com.money.exception.ErrorCode;
+import com.money.exception.MainException;
+
+public class OverFlowInviteCodeCountException extends MainException {
+
+    public OverFlowInviteCodeCountException() {
+        super(ErrorCode.OVER_FLOW_INVITE_CODE);
+    }
+}
