+++ subprojects/domain/src/main/java/com/money/domain/member/util/FollowStatus.java
@@ -1,0 +1,5 @@
+package com.money.domain.member.util;
+
+public enum FollowStatus {
+    HOLD, ACCEPTED
+}
