--- subprojects/member/src/main/java/com/money/exception/ExpiredCodeException.java
@@ -1,8 +1,0 @@
-package com.money.exception;
-
-public class ExpiredCodeException extends MainException{
-
-    public ExpiredCodeException() {
-        super(ErrorCode.EXPIRED_CODE);
-    }
-}
