--- subprojects/member/src/main/java/com/money/exception/MemberAlreadyExistException.java
@@ -1,8 +1,0 @@
-package com.money.exception;
-
-public class MemberAlreadyExistException extends MainException{
-
-    public MemberAlreadyExistException() {
-        super(ErrorCode.ALREADY_EXIST_MEMBER);
-    }
-}
