--- subprojects/member/src/main/java/com/money/exception/NotFoundMemberException.java
@@ -1,8 +1,0 @@
-package com.money.exception;
-
-public class NotFoundMemberException extends MainException{
-
-    public NotFoundMemberException() {
-        super(ErrorCode.NOT_FOUND_MEMBER);
-    }
-}
