--- subprojects/member/src/main/java/com/money/exception/OverFlowInviteCodeCountException.java
@@ -1,8 +1,0 @@
-package com.money.exception;
-
-public class OverFlowInviteCodeCountException extends MainException{
-
-    public OverFlowInviteCodeCountException() {
-        super(ErrorCode.OVER_FLOW_INVITE_CODE);
-    }
-}
