--- subprojects/member/src/main/java/com/money/util/FollowStatus.java
@@ -1,5 +1,0 @@
-package com.money.util;
-
-public enum FollowStatus {
-    HOLD, ACCEPTED
-}
