--- subprojects/payment/build.gradle
@@ -1,23 +1,0 @@
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    compileOnly 'org.projectlombok:lombok'
-    runtimeOnly 'com.h2database:h2'
-    runtimeOnly 'com.mysql:mysql-connector-j'
-    annotationProcessor 'org.projectlombok:lombok'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-
-    api project(':common')
-    api project(':member')
-}
-
-tasks.named('test') {
-    useJUnitPlatform()
-}
-
-tasks.named('bootJar') {
-    enabled = false
-}
-
-tasks.named('jar') {
-    enabled = true
-}
